Refactor the import statements logic to make it more readable - Moved the logic to a new helper module import-helper.js - Refactored the bulky functions and updated comments
